requirements.txt
-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10; python_version >= '3'
cryptography==36.0.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.47.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-filter==21.1
django-qr-code==3.0.0
django-rest-auth[with-social]==0.9.5
django-rest-framework==0.1.0
django-verify-email==1.0.9
django==3.2.7
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangorestframework==3.13.1
drf-extra-fields==3.2.1
drf-flex-fields==0.9.7
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==3.3; python_version >= '3'
oauthlib==3.1.1; python_version >= '3.6'
pillow==9.0.0
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==1.7.1
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
requests-oauthlib==1.3.0
requests==2.27.1; python_version >= '3.0'
segno==1.4.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
stripe==2.64.0
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.3.0
